Refactor sharing of Transfer task submit options
Move these all into a dedicated module, and make sure more are shared
between `globus transfer` and `globus timer create transfer`.

In cases where the timer option name appeared to be more correct in
some way (e.g. `--encrypt-data` over `--encrypt`, which matches the
API fieldname), both names are preserved for the `globus transfer`
command only, by means of aliases.
